Create 1_CreateTrainingData
This codes sets up Google Earth Engine to make creating training data a little simpler. We use an unsupervised classification to get an idea of how the computer sees the Landsat imagery. This should be your first step in classifying landcover via the model we used for my thesis work.